Switched to sections instead of pages, moved several CSS classes to global styles and set to nly one page called Home that imports all other sections
