--- src/pages/work/Work.css
@@ -1,41 +1,0 @@
-.work-container {
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    background-color: white;
-}
-
-.work-content {
-    flex: 1;
-    padding: 2rem 15%;
-    display: flex;
-    flex-direction: column;
-}
-
-/* Remove the old .work-title styles since we're using .section-title now */
-
-/* Add these styles if they're not already defined globally */
-.section-title {
-    font-size: 2.5rem;
-    font-weight: 600;
-    margin-bottom: 3rem;
-    position: relative;
-    display: inline-block;
-}
-
-.section-title::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    bottom: -10px;
-    width: 60px;
-    height: 3px;
-    background-color: #0984e3;
-}
-
-/* Add the same media query as Skills component for consistency */
-@media (max-width: 768px) {
-    .work-content {
-        padding: 4rem 5%;
-    }
-}
