+++ src/sections/about/About.css
@@ -1,38 +1,12 @@
-.about-container {
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    background-color: white;
-}
-
-.about-content {
-    flex: 1;
-    padding: 4rem 10% 2rem;
-}
-
 .about-section {
+    padding: 2rem 10%;
+    background-color: white;
+    position: relative;
     display: flex;
     gap: 4rem;
     max-width: 1200px;
     margin: 0 auto;
     align-items: flex-start;
-}
-
-.section-title {
-    font-size: 2.5rem;
-    margin-bottom: 3rem;
-    position: relative;
-    display: inline-block;
-}
-
-.section-title::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    bottom: -10px;
-    width: 60px;
-    height: 3px;
-    background-color: #0984e3;
 }
 
 .about-image {
