+++ src/sections/skills/Skills.css
@@ -1,27 +1,3 @@
-.skills-section {
-    padding: 2rem 15%;
-    background-color: white;
-    position: relative;
-}
-
-/* TODO: Cleanup this css and add to global styles/separate component instead */
-.section-title {
-    font-size: 2.5rem;
-    margin-bottom: 3rem;
-    position: relative;
-    display: inline-block;
-}
-
-.section-title::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    bottom: -10px;
-    width: 60px;
-    height: 3px;
-    background-color: #0984e3;
-}
-
 .skills-grid {
     display: grid;
     grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
@@ -68,11 +44,6 @@
 }
 
 @media (max-width: 768px) {
-    .skills-section {
-        padding: 4rem 5%;
-        overflow: hidden;
-    }
-
     .skills-grid {
         display: flex;
         overflow-x: auto;
