+++ src/styles/global.css
@@ -1,0 +1,82 @@
+/* Section */
+.section {
+    padding: 2rem 10%;
+    background-color: white;
+    position: relative;
+}
+
+/* Section Titles */
+.section-title {
+    font-size: 2.5rem;
+    margin-bottom: 3rem;
+    position: relative;
+    display: inline-block;
+}
+
+.section-title::after {
+    content: '';
+    position: absolute;
+    left: 0;
+    bottom: -10px;
+    width: 60px;
+    height: 3px;
+    background-color: #0984e3;
+}
+
+/* Buttons */
+.primary-button, .secondary-button {
+    padding: 0.8rem 2rem;
+    border-radius: 4px;
+    font-weight: 500;
+    text-decoration: none;
+    transition: all 0.3s ease;
+    min-width: 120px;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+}
+
+.primary-button {
+    background-color: #0984e3;
+    color: white;
+}
+
+.primary-button:hover {
+    background-color: #0773c5;
+    transform: translateY(-2px);
+}
+
+.secondary-button {
+    border: 2px solid #0984e3;
+    color: #0984e3;
+    gap: 8px;
+}
+
+.secondary-button:hover {
+    background-color: #0984e3;
+    color: white;
+    transform: translateY(-2px);
+}
+
+@media (max-width: 768px) {
+    /* Section */
+    .section {
+        padding: 4rem 5%;
+        overflow: hidden;
+    }
+
+    /* Buttons */
+    .primary-button, .secondary-button {
+        min-width: 140px;
+        padding: 0.7rem 1.5rem;
+        font-size: 0.9rem;
+    }
+}
+
+@media (max-width: 480px) {
+    .primary-button, .secondary-button {
+        width: 100%;
+        min-width: unset;
+    }
+}
